Refs #8641. Moved blackbox test code to subfolder
