+++ test/blackbox/common/RTPSWithRegistrationReader.hpp
