+++ test/blackbox/common/RTPSWithRegistrationWriter.hpp
