+++ test/blackbox/utils/data_generators.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "../BlackboxTests.hpp"
+#include "../common/BlackboxTests.hpp"
 
 #include <cstddef>
 #include <cstdint>
