+++ test/blackbox/utils/print_functions.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "../BlackboxTests.hpp"
+#include "../common/BlackboxTests.hpp"
 
 #include <iostream>
